Use '.jsx' extension instead of '.js'; Use `import * as React from 'react'`
