+++ examples/async/src/components/Picker.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import * as React from 'react'
 import PropTypes from 'prop-types'
 
 export default class Picker extends React.Component {
