+++ examples/async/src/components/Posts.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import * as React from 'react'
 import PropTypes from 'prop-types'
 
 export default class Posts extends React.Component {
