+++ examples/async/src/containers/App.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import * as React from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import { selectReddit, invalidateReddit } from '../actions'
