+++ examples/async/src/main.js
@@ -1,5 +1,5 @@
-import "@babel/polyfill"
-import React from 'react'
+import '@babel/polyfill'
+import * as React from 'react'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
 import App from './containers/App'
@@ -13,5 +13,5 @@ render(
   <Provider store={store}>
     <App />
   </Provider>,
-  document.getElementById('root')
+  document.getElementById('root'),
 )
