+++ examples/counter/src/main.js
@@ -1,12 +1,12 @@
 import '@babel/polyfill'
 
-import React from 'react'
+import * as React from 'react'
 import ReactDOM from 'react-dom'
 import { createStore, applyMiddleware } from 'redux'
 import createSagaMiddleware from 'redux-saga'
 import sagaMonitor from '../../sagaMonitor'
 
-import ErrorGenerator from './components/ErrorGenerator'
+import Counter from './components/Counter'
 import reducer from './reducers'
 import rootSaga from './sagas'
 
@@ -17,7 +17,16 @@ sagaMiddleware.run(rootSaga)
 const action = type => store.dispatch({ type })
 
 function render() {
-  ReactDOM.render(<ErrorGenerator value={store.getState()} action={action} />, document.getElementById('root'))
+  ReactDOM.render(
+    <Counter
+      value={store.getState()}
+      onIncrement={() => action('INCREMENT')}
+      onDecrement={() => action('DECREMENT')}
+      onIncrementIfOdd={() => action('INCREMENT_IF_ODD')}
+      onIncrementAsync={() => action('INCREMENT_ASYNC')}
+    />,
+    document.getElementById('root'),
+  )
 }
 
 render()
