+++ examples/real-world/containers/DevTools.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import * as React from 'react'
 import { createDevTools } from 'redux-devtools'
 import LogMonitor from 'redux-devtools-log-monitor'
 import DockMonitor from 'redux-devtools-dock-monitor'
