+++ examples/real-world/index.js
@@ -1,6 +1,6 @@
 import '@babel/polyfill'
 // React imports
-import React from 'react'
+import * as React from 'react'
 import { render } from 'react-dom'
 
 // app specific imports
