+++ examples/real-world/routes.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import * as React from 'react'
 import { Route } from 'react-router'
 import App from './containers/App'
 import UserPage from './containers/UserPage'
