+++ examples/real-world/server.js
@@ -8,7 +8,7 @@ import express from 'express'
 import path from 'path'
 import favicon from 'serve-favicon'
 import routes from './routes'
-import React from 'react'
+import * as React from 'react'
 import Root from './containers/Root'
 import { renderToString } from 'react-dom/server'
 import { match, createMemoryHistory } from 'react-router'
