+++ examples/shopping-cart/src/components/Cart.js
