+++ examples/shopping-cart/src/components/CartItem.js
