+++ examples/shopping-cart/src/components/Product.js
