+++ examples/shopping-cart/src/components/ProductItem.js
